chore: code structure updated
