+++ utils/model.py
@@ -1,0 +1,5 @@
+users: list[dict] = [
+    {'name': 'Emilia', 'location': 'Lublin', 'posts': 500, },
+    {'name': 'Julia', 'location': 'Warszawa', 'posts': 1000, },
+    {'name': 'Martyna', 'location': 'Ząbki', 'posts': 200, }
+]
